Bump version to  beta 2. Allow caller to unset/remove min width/height parameters from an an server request. Cleanup typos/grammar in some error messages. When ad debug logging is enabled, automatically write log messages to a file on the sdcard (with default name: mOcean-sample-log.txt). Caller can still override the file name with the set file log call. Remove dead advertiser id / group code. Disable sensor usage in ormma sensor controller when on emulator, as this sometimes causes hangs.

+++ trunk/Sources/src/com/MASTAdView/MASTAdLog.java
@@ -7,6 +7,7 @@ import android.app.ActivityManager;
 import android.app.Application;
 import android.content.Context;
 import android.content.pm.PackageManager;
+import android.os.Environment;
 import android.util.Log;
 
 public class MASTAdLog {
@@ -21,7 +22,8 @@ public class MASTAdLog {
 	
 	int CurrentLogLevel = 0;
 	String AppName="";
-	
+	private static String defaultLogFileName =  Environment.getExternalStorageDirectory().getAbsolutePath() + "/mOcean-sample-log.txt";
+	private static boolean loggingToFile = false;
 	MASTAdViewCore adView;
 	
 	private static int DefaultLevel = LOG_LEVEL_NONE;
@@ -29,16 +31,28 @@ public class MASTAdLog {
 	public static void setDefaultLogLevel(int logLevel)
 	{
 		DefaultLevel = logLevel;
+		if ((logLevel > 0) && (loggingToFile == false))
+		{
+			setFileLog(defaultLogFileName); // log to default file
+		}
 	}
 	
 	public static void setFileLog(String fileName)
 	{
+		if ((fileName == null) || (fileName.length() < 1))
+		{
+			return;
+		}
+		
 		try {     
 			File filename = new File(fileName);
 			if (filename.exists()) filename.delete();
 			filename.createNewFile();
-			String cmd = "logcat -v time -f "+filename.getAbsolutePath();
-			Runtime.getRuntime().exec(cmd); 
+			String cmd = "logcat -v time -f "+filename.getAbsolutePath(); // + " com.MASTAdView.";
+			Runtime.getRuntime().exec(cmd);
+			loggingToFile = true;
+			//log(LOG_LEVEL_1,LOG_TYPE_INFO,"SetFileLog","Logging to file: " + fileName);
+			System.out.println("Logging to file: " + fileName);
 		} catch (IOException e) { 
 				e.printStackTrace(); 
 		}
@@ -76,6 +90,11 @@ public class MASTAdLog {
 		case LOG_LEVEL_3:log(LOG_LEVEL_1,LOG_TYPE_INFO,"SetLogLevel","LOG_LEVEL_3");break;
 		default:
 			log(LOG_LEVEL_1,LOG_TYPE_INFO,"SetLogLevel","LOG_LEVEL_NONE");
+		}
+		
+		if ((logLevel > 0) && (loggingToFile == false))
+		{
+			setFileLog(defaultLogFileName); // log to default file
 		}
 	}
 }
